Fixes padding bug.
Explanation: If the frame width or height is less than the inset,
the width or height of the frame can evalute to infinity, which
results in bad rendering and crashes on zoom.

The fix ensures that this is checked before applying the padding.

Because the check is done for horizontal and vertical inset
seperately, the content mode of the view is explicitly set.